+++ CHANGES.rst
@@ -7,6 +7,8 @@ mio 0.0.5.dev
 
 - Split up core into core and types.
 - Re-implemented ``True``, ``False`` and ``None`` in mio.
+- Implemented ``bin`` builtin.
+- Implemented ``bool`` builtin.
 
 
 mio 0.0.4 (*2013-10-27*)